finalizei o cabeçalho da pagina principal e realizei a troca de cabeçalho para as outras paginas
