+++ src/app/app.component.html
@@ -1,3 +1,4 @@
-<app-header></app-header>
+<app-header *ngIf="isMainPage"></app-header>
+<app-other-header *ngIf="!isMainPage"></app-other-header>
 <router-outlet></router-outlet>
 <app-footer></app-footer>
