+++ src/app/components/other-header/other-header.component.html
@@ -1,0 +1,30 @@
+<header id="header">
+    <nav id="nav">
+      <a class="brand" href="/">
+        <img id="logo" class="logo" src="/logo.svg" alt="Logo Branca" />
+      </a>
+  
+      <ul class="links">
+        <li><a href="/about">Sobre</a></li>
+        <li><a href="/services">Serviços</a></li>
+        <li><a href="/carreiras">Carreiras</a></li>
+        <li><a href="/contact">Contato</a></li>
+      </ul>
+    </nav>
+  
+    <div class="mobile-menu-icon">
+      <button onClick="menuShow()">
+        <img class="icon" src="/Menu.svg" alt="" />
+      </button>
+    </div>
+  
+    <div class="mobile-menu">
+      <ul class="links">
+        <li><a href="/about">Sobre</a></li>
+        <li><a href="/services">Serviços</a></li>
+        <li><a href="/career">Carreiras</a></li>
+        <li><a href="/contact">Contato</a></li>
+      </ul>
+    </div>
+  </header>
+
